project/accountProject/backend/requirements.txt
asgiref==3.4.1
attrs==21.2.0
backports.entry-points-selectable==1.1.0
bcrypt==3.2.0
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.6
cryptography==3.4.8
defusedxml==0.7.1
distlib==0.3.2
dj-rest-auth==2.1.11
Django==3.2.7
django-allauth==0.45.0
django-cors-headers==3.10.1
django-extensions==3.1.3
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
filelock==3.0.12
idna==3.2
iniconfig==1.1.1
oauthlib==3.1.1
packaging==21.3
pbr==5.6.0
pendulum==2.1.2
platformdirs==2.2.0
pluggy==1.0.0
py==1.11.0
pycparser==2.20
PyJWT==2.1.0
pyparsing==3.0.6
pytest==6.2.5
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.1
pytzdata==2020.1
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
sqlparse==0.4.2
stevedore==3.3.0
toml==0.10.2
urllib3==1.26.7
virtualenv==20.6.0
virtualenv-clone==0.5.6
virtualenvwrapper==4.8.4
